backend/requirements.txt
fastapi
uvicorn
pydantic-settings
supabase
playwright
httpx
beautifulsoup4
apscheduler
openai
anthropic
google-generativeai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
resend
